+++ operator/CHANGELOG.md
@@ -12,4 +12,10 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 ### Added
 
+- Support for the following env vars for testing 
+  - AIS_TEST_NODE_IMAGE
+  - AIS_TEST_PREV_NODE_IMAGE
+  - AIS_TEST_INIT_IMAGE
+  - AIS_TEST_PREV_INIT_IMAGE
+
 ### Changed